Move Z-Wave config API to config.zwave
